change type to f64

+++ src/activations.rs
@@ -32,18 +32,18 @@ pub fn string_to_activation(val:&str) -> Activations {
 	}
 }
 
-pub fn sigmoid(x:f32) -> f32
+pub fn sigmoid(x:f64) -> f64
 {
 	let k = x.exp();
 	k / (1.0 + k)
 }
 
-pub fn tanh(x:f32) -> f32
+pub fn tanh(x:f64) -> f64
 {
 	x.tanh()
 }
 
-pub fn relu(x:f32) -> f32
+pub fn relu(x:f64) -> f64
 {
 	match 0.0 >= x {
 		true => {0.0},
@@ -51,7 +51,7 @@ pub fn relu(x:f32) -> f32
 	}
 }
 
-pub fn leakyrelu(x:f32) -> f32
+pub fn leakyrelu(x:f64) -> f64
 {
 	match 0.0 >= x {
 		true => {0.01 * x},
@@ -59,17 +59,17 @@ pub fn leakyrelu(x:f32) -> f32
 	}
 }
 
-pub fn sigmoid_deriv(x:f32) -> f32
+pub fn sigmoid_deriv(x:f64) -> f64
 {
 	x * (1.0 - x)
 }
 
-pub fn tanh_deriv(x:f32) -> f32
+pub fn tanh_deriv(x:f64) -> f64
 {
 	1.0 - (x * x)
 }
 
-pub fn relu_deriv(x:f32) -> f32
+pub fn relu_deriv(x:f64) -> f64
 {
 	match 0.0 >= x {
 		true => {0.0},
@@ -77,7 +77,7 @@ pub fn relu_deriv(x:f32) -> f32
 	}
 }
 
-pub fn leakyrelu_deriv(x:f32) -> f32
+pub fn leakyrelu_deriv(x:f64) -> f64
 {
 	match 0.0 >= x {
 		true => {0.01},
@@ -85,7 +85,7 @@ pub fn leakyrelu_deriv(x:f32) -> f32
 	}
 }
 
-pub fn softmax(x:f32) -> f32
+pub fn softmax(x:f64) -> f64
 {
 	x.exp()
 }
